2021 05 06 14:05 커밋(장바구니)

+++ SemiProject_team01/src/cart/model/CartDAO.java
@@ -1,5 +1,7 @@
 package cart.model;
 
+import java.io.UnsupportedEncodingException;
+import java.security.GeneralSecurityException;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -13,7 +15,7 @@ import javax.naming.InitialContext;
 import javax.naming.NamingException;
 import javax.sql.DataSource;
 
-
+import member.model.MemberVO;
 import product.model.*;
 
 public class CartDAO implements InterCartDAO {
@@ -241,25 +243,39 @@ public class CartDAO implements InterCartDAO {
 		try {
 			conn = ds.getConnection();
 			
-			String  sql = " insert into tbl_wishlist(wnum, fk_userid, fk_pnum, oqty , fk_pdetailnum) " + 
-					" values (seq_wish_wnum.nextval, ? , ?, 1, ?)";
-       
+			String sql = " select wnum, fk_userid, fk_pnum, oqty , fk_pdetailnum "+
+			             " from tbl_wishlist "+
+					     " where fk_userid = ? and fk_pdetailnum = ?";
+			
 			pstmt = conn.prepareStatement(sql);
 			pstmt.setString(1, paraMap.get("userid"));
-			pstmt.setString(2, paraMap.get("fk_pnum")); // userid가 들어오면 2로 수정
-			pstmt.setString(3, paraMap.get("fk_pdetailnum")); // userid가 들어오면 4로 수정
+			pstmt.setString(2, paraMap.get("fk_pdetailnum"));
 			
-			pstmt.executeUpdate();
+			rs = pstmt.executeQuery();
 			
-			sql = " delete from tbl_cart "+
-						 " where fk_userid='user' and cartnum= ? ";
-			pstmt = conn.prepareStatement(sql);
-			
-			pstmt.setString(1, paraMap.get("userid"));
-			pstmt.setString(1, paraMap.get("cartnum")); // userid가 들어오면 2로 수정
-			
-			n= pstmt.executeUpdate();
-			
+			if(rs.next()) {
+				n=0;
+			}	
+			else {
+				sql = " insert into tbl_wishlist(wnum, fk_userid, fk_pnum, oqty , fk_pdetailnum) " + 
+						" values (seq_wish_wnum.nextval, ? , ?, 1, ?)";
+	       
+				pstmt = conn.prepareStatement(sql);
+				pstmt.setString(1, paraMap.get("userid"));
+				pstmt.setString(2, paraMap.get("fk_pnum")); // userid가 들어오면 2로 수정
+				pstmt.setString(3, paraMap.get("fk_pdetailnum")); // userid가 들어오면 4로 수정
+				
+				pstmt.executeUpdate();
+				
+				sql = " delete from tbl_cart "+
+							 " where fk_userid= ? and cartnum= ? ";
+				pstmt = conn.prepareStatement(sql);
+				
+				pstmt.setString(1, paraMap.get("userid"));
+				pstmt.setString(2, paraMap.get("cartnum")); // userid가 들어오면 2로 수정
+				
+				n= pstmt.executeUpdate();
+			}
 			
 		} finally {
 			close();
@@ -268,44 +284,9 @@ public class CartDAO implements InterCartDAO {
 		return n;
 	}
 
-	/*
-	// 해당 상품에 대한 옵션 리스트 보이기(select)
-	@Override
-	public List<String> optionSelect(String fk_pnum) throws SQLException {
-
-		List<String> optionList =  new ArrayList<>();
-		
-		try {
-			conn = ds.getConnection();
-			
-			String sql = " select D.optionname "+
-					" from tbl_proddetail D  "+
-					" join "+
-					" tbl_product P "+
-					" on D.fk_pnum = P.pnum "+
-					" where D.fk_pnum= ? ";
-			
-			pstmt = conn.prepareStatement(sql);
-			pstmt.setString(1, fk_pnum);
-		
-			
-			rs = pstmt.executeQuery();
-		
-			while(rs.next()) {
-				
-				String option = rs.getString(1);
-				optionList.add(option);
-			}
-		
-		}finally {
-			close();
-		}
-		
-		return optionList;
-	}
-*/
 	
-	// 해당 상품에 대한 옵션 리스트 보이기(select) 2
+	
+	// 해당 상품에 대한 옵션 리스트 보이기(select) 
 	@Override
 	public List<ProductDetailVO> optionSelect(String fk_pnum) throws SQLException{
 
@@ -583,15 +564,172 @@ public class CartDAO implements InterCartDAO {
 	}
 
 	
+	// 해당 상품을 위시리스트에 삭제(delete)하고 장바구니에 추가(insert)
+	@Override
+	public int deleteWishAddCart(Map<String, String> paraMap) throws SQLException {
+		
+		int n=0;
+		
+		try {
+			conn = ds.getConnection();
+			
+			String sql = " select cartnum, fk_userid, fk_pnum, oqty , fk_pdetailnum "+
+			             " from tbl_cart "+
+					     " where fk_userid = ? and fk_pdetailnum = ?";
+			
+			pstmt = conn.prepareStatement(sql);
+			pstmt.setString(1, paraMap.get("userid"));
+			pstmt.setString(2, paraMap.get("fk_pdetailnum"));
+			
+			rs = pstmt.executeQuery();
+			
+			if(rs.next()) {
+				n=0;
+			}	
+			else {
+				sql = " insert into tbl_cart(cartnum, fk_userid, fk_pnum, oqty , fk_pdetailnum) " + 
+						" values (seq_cart_num.nextval, ? , ?, 1, ?)";
+	       
+				pstmt = conn.prepareStatement(sql);
+				pstmt.setString(1, paraMap.get("userid"));
+				pstmt.setString(2, paraMap.get("fk_pnum")); 
+				pstmt.setString(3, paraMap.get("fk_pdetailnum")); 
+				
+				pstmt.executeUpdate();
+				
+				sql = " delete from tbl_wishlist "+
+							 " where fk_userid= ? and wnum= ? ";
+				pstmt = conn.prepareStatement(sql);
+				
+				pstmt.setString(1, paraMap.get("userid"));
+				pstmt.setString(2, paraMap.get("wnum")); 
+				
+				n= pstmt.executeUpdate();
+			}
+			
+		} finally {
+			close();
+		}
+				
+		return n;
+	}
 
 	
 	
+	// 페이징처리를 위해서 회원의 위시리스트 총페이지 개수 알아오기(select)  
+	@Override
+	public int selectTotalPage(Map<String, String> paraMap) throws SQLException {
+		
+		
+			int totalPage=0;
+			
+			try {
+				conn=ds.getConnection();
+				
+				String sql = " select ceil(count(*) / ?) "
+						   + " from tbl_wishlist"
+						   + " where fk_userid = ? ";
+				
+				pstmt = conn.prepareStatement(sql);
+				pstmt.setString(1, paraMap.get("sizePerPage"));
+				pstmt.setString(2, paraMap.get("userid"));
+				
+				
+				rs = pstmt.executeQuery();
+				rs.next();
+				
+				totalPage = rs.getInt(1);
+		
+			} finally {
+				close();
+			}
+			return totalPage;
+		
+
+	}
+
 	
-	
-	
-	
-	
-	
-	
+	@Override
+	public List<WishListVO> selectPagingMember(Map<String, String> paraMap) throws SQLException {
+
+			List<WishListVO> wishPageList = new ArrayList<>();
+			
+			try {
+				conn = ds.getConnection();
+				
+				String sql =" select W.fk_pnum, P.pimage1,  P.pname, W.oqty, D.optionname, P.saleprice, W.fk_pdetailnum, W.wnum " + 
+						"from(" + 
+						"    select rno, wnum, fk_userid, fk_pnum, oqty,fk_pdetailnum " + 
+						"    from( " + 
+						"        select rownum as rno, wnum, fk_userid, fk_pnum, oqty,fk_pdetailnum " + 
+						"        from " + 
+						"        (select wnum, fk_userid, fk_pnum, oqty,fk_pdetailnum  " + 
+						"        from tbl_wishlist " + 
+						"        where fk_userid = ?  " + 
+						"        order by wnum desc " + 
+						"        )V " + 
+						"    )T " + 
+						"    )W join tbl_product P " + 
+						" on W.fk_pnum = P.pnum  " + 
+						" join  " + 
+						" tbl_proddetail D  " + 
+						" on W.fk_pdetailnum = D.pdetailnum  " + 
+						" where rno between ? and ? ";
+				
+				
+			
+				//	(currentShowPageNo * sizePerPage) - (sizePerPage - 1) and (currentShowPageNo * sizePerPage)
+				
+				int currentShowPageNo = Integer.parseInt(paraMap.get("currentShowPageNo"));
+				int sizePerPage = Integer.parseInt(paraMap.get("sizePerPage"));
+				
+				pstmt = conn.prepareStatement(sql);
+
+					pstmt.setString(1, paraMap.get("userid"));
+					pstmt.setInt(2, (currentShowPageNo * sizePerPage) - (sizePerPage - 1));
+					pstmt.setInt(3, (currentShowPageNo * sizePerPage));
+				
+				rs = pstmt.executeQuery();
+				
+				while(rs.next()) {
+					
+					int fk_pnum = rs.getInt(1);
+					String pimage1 = rs.getString(2);
+					String pname= rs.getString(3);
+					int oqty = rs.getInt(4);
+					String optionname = rs.getString(5);
+					int saleprice = rs.getInt(6);
+					int fk_pdetailnum = rs.getInt(7);
+					int wnum = rs.getInt(8);
+					
+					WishListVO wvo = new WishListVO();  
+					wvo.setFk_pnum(fk_pnum);
+					wvo.setOqty(oqty);
+					wvo.setFk_pdetailnum(fk_pdetailnum);
+					wvo.setWnum(wnum);
+					
+					ProductVO pvo = new ProductVO();
+					pvo.setPimage1(pimage1);
+					pvo.setPname(pname);
+					pvo.setSaleprice(saleprice);
+					
+					ProductDetailVO pdetailvo = new ProductDetailVO();
+					pdetailvo.setOptionname(optionname);
+					
+					wvo.setPvo(pvo);
+					wvo.setPdetailvo(pdetailvo);
+					
+					wishPageList.add(wvo);
+				}
+				
+		
+			} finally {
+				close();
+			}
+
+			return wishPageList;
+		}
+
+
 
 }
